+++ CHANGELOG.md
@@ -74,6 +74,7 @@
 - [9204](https://github.com/vegaprotocol/vega/issues/9204) - Ensure teams names are not duplicates
 - [9080](https://github.com/vegaprotocol/vega/issues/9080) - Add support for vested and vesting account in GraphQL
 - [9147](https://github.com/vegaprotocol/vega/issues/9147) - Add reward multiplier to vesting engine
+- [9234](https://github.com/vegaprotocol/vega/issues/9234) - Add support for transfers out of the vested account
 
 
 ### 🐛 Fixes